CHROMIUM: fix incorrect crossystem recovery_reason
U-Boot should not parse the raw contents of VbNvStorage, and so cannot
read the recovery reason from the VbNvStorage. On the other hand, it is
easy for crossystem to read the recovery reason from the VbNvStorage
itself.

After this change is merged, U-Boot will stop providing the (incorrect)
recovery reason in the device tree.

BUG=chromium-os:17876,chromium-os:17852
TEST=press recovery button and see crossystem reports recovery_reason=2

Change-Id: I236667f0b4f2e25da193cf6b6f7db3871d1e093f