Update patch set 1
Patch Set 1:

gimp is too huge with lots of dependency that may bloat the chroot, so I think we probably won't want to pull that into chroot.

ImageMagick is small & powerful - except that it seems lack of I18N font processing? I wonder if PIL can do that for us.

Well, these are long term plans. Since we want to deliver the image for ARM PVT & MP, I suggest let's use Bill's CL first and improve this in future.

Patch-set: 1