Update patch set 1
Patch Set 1:

I wouldn't bother trying to make this work inside the chroot. There are too many dependencies and they serve little purpose.  We only need to do this once, until/unless there's a change to the text, then we'll just check in the generated images and use those as the master.

Patch-set: 1