Update patch set 2
Patch Set 2:

commented a bit more on the thread.  Still not sold on the justification, but if you decide to land this, I'd appreciate an extension to chromeos-boot-alert to go with it.  (Or your review of said change at the very least!)

thanks!
will

Patch-set: 2
Reviewer: Gerrit User 1000019 <1000019@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0