Make crossystem on x86 understand booting from the read only firmware
Without this change, when crossystem tries to interpret the constant describing
the main active firmware in VbGetArchPropertyString, it uses the default case
and returns NULL to the caller. The caller then attempts to look up the
firmware index in the VDAT area, and because that setting has no way to
represent using the readonly firmware directly, it incorrectly decides that the
read/write firmware A was used.

By making the switch statement in VbGetArchPropertyString return a valid
string, crossystem gets the right answer without ever seeing the misleading
value in vdat.

BUG=chrome-os-partner:5944
BUG=chrome-os-partner:5961
BUG=chrome-os-partner:5962
TEST=Booted with this crossystem on coreboot/u-boot with a change that
populates the field describing the current main firmware with real data.
Saw that crossystem reported that the current main firmware was RO which is
correct.

Change-Id: I002517028ab0c0c35c41bd16154aa1cd57960938
Signed-off-by: Gabe Black <gabeblack@google.com>