Update patch set 5
Patch Set 5:

To Randall's inline comment, the link to the bug was removed from the code since it is on the partner tracker. 

I agree with the idea from Luigi but I think the value we check for needs to be TPM_E_IOERROR, based on the log from the bug, SetupTPM() is returning 0x1F, which appears to correspond to TPM_E_IOERROR. Does this value also make sense or am I confused?

Patch-set: 5