Update patch set 1
Patch Set 1: (2 inline comments)

re: devkeys. I was thinking about that. I don't think the default builders touch the firmware at all (i.e. they don't try to re-sign). So it is whatever was checked in to the private repository. I am guessing the build process generates the dev-key signed firmware with the right set of keyblocks. It may be useful to have 2 versions of devkeys - one for use on Alex/ZGB, and one for others.

Patch-set: 1