Add in a platform_family value to crossystem
This is an initial pass at adding in a platform_family value to crossystem,
as the platform (particularly for ARM) is not easily accessable elsewhere at
runtime.

DO NOT SUBMIT yet, this is just an initial concept and needs work, this is
for discussion purposes only at the moment.

*Need to decide if this addition makes sense to have, and if 'platform_family'
is the right word.
*Need to decide if the 'model name' string from /proc/cpuinfo is suitable as
a 'platform_family' value for the x86, this probably needs simplified, as the
full description with speed may not be necessary.
*Need to further test the ARM side and possibly add a FDT entry specific to
the platform value (possibly pending this on another CL), currently the ARM
side is not functioning properly due to this (see comment).

BUG=chromium-os:24669
TEST=Manual

Change-Id: Id0e973902d27ead471c1243bcc6c3292acc8479d