Add in a platform value to crossystem
This is an initial pass at adding in a platform value to crossystem, as the
platform (particularly for ARM) is not easily accessable elsewhere at runtime.

DO NOT SUBMIT yet, this is just an initial concept and needs work, this is
for discussion purposes only at the moment.

*Need to decide if this addition makes sense to have (maybe this belongs in
mosys?), and if 'platform' is the right word.
*Need to decide if the 'model name' string from /proc/cpuinfo is suitable as
a 'platform' value for the x86.
*Need to further test the ARM side and possibly add a FDT entry specific to
the platform value (possibly pending this on another CL), currently the ARM
side is not functioning properly due to this (see comment).

BUG=None
TEST=Manual

Change-Id: Id0e973902d27ead471c1243bcc6c3292acc8479d