Add in a platform_family value to crossystem
This is an initial pass at adding in a platform_family value to crossystem,
as the platform (particularly for ARM) is not easily accessable elsewhere at
runtime.

DO NOT SUBMIT yet, this is just an initial concept and needs work, this is
for discussion purposes only at the moment.

For the ARM side this contains a table which is used to determine the platform
family based on the /proc/device-tree/compatible entry. Similarly on x86 the
table is used to check against PCI entries. Additional entries can be made
as new platform families emerge.

TODO:
*Verify nomenclature for the tables.

BUG=chromium-os:24669
TEST=Manual

Change-Id: Id0e973902d27ead471c1243bcc6c3292acc8479d