Update patch set 1
Patch Set 1: (3 inline comments)

Thank you for the feedback, if this goes further I can file a bug on it with some more detail as you suggest, but for now I will try to clear some confusion here. The problem this is looking to bring up and solve is that there is not a good single point to determine what actual processor is in use. This stemmed from the potential need to do things differently (such as different flags being passed to Chrome) depending on what the 'platform' is. Perhaps 'platform' is the wrong word, since it can be a bit ambiguous, so I am open to suggestion, ultimately I would like to see this return some indication of the actual chipset or SoC being used (Atom xxx vs. Pentium xxx vs. Tegra xxx, etc), something more specific than architecture.

Patch-set: 1