Update patch set 1
Patch Set 1:

Talk to tbroch, he was writing some (x86) code that was doing something similar. So having a generic way would be nice. But I suspect that to be useful it will return something that is less specific than the model name, but more specific than the architecture. You would want to be able to tell a 1st generation i5 from a second generation (Sandy Bridge) i5 but you probably wouldn't care which of the frequency bins it is.

Patch-set: 1