Update patch set 1
Patch Set 1:

It's hard to tell in gerrit, but if this is a cherry pick, it should be deleting the stub implementation and making a symlink to the existing 32bit implementation.  Is that what you see in your repository with this change?

Patch-set: 1
Reviewer: Gerrit User 1000153 <1000153@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=0