Build firmware vboot code in 32-bit mode on 64-bit platforms
On 64-bit overlays, all the firmware related code is still compiled in
32-bit mode and we want  the same flags as i386 mode.
But as we have a 64-bit toolchain, we need to add "-m32" flag to force the
build in 32-bit mode.

BUG=chrome-os-partner:7568
TEST=emerge-[lumpy64|stumpy|tegra2_kaen] vboot_reference-firmware
along with the other related changes, build a firmware in the lumpy64
overlay and test it on a lumpy machine.

Change-Id: I40dd8643cb114c8090fa35a7e0aa55ca5e1f2877