During the dev screen, an additional F3 key to trigger dev USB boot.
Due to the limitation of servo that is unable to send U keys, dev USB boot
(triggered by Ctrl-U) is unable to be tested on FAFT. To solve it, firmware
should add an addition key combination to workaround it. F3 key is the one.

BUG=chrome-os-partner:6759
TEST=compile the firmware and update it to Lumpy; during the dev screen,
press F3 to trigger USB boot.

Change-Id: If146d6680b68e21e18bbd6324bfcfd4f19adcff2