Update patch set 6
Patch Set 6: (25 inline comments)

The -fPIC option was added based on the suggestion given by G++ on amd64, without it amd64 won't build the CgptManagerTests binary that uses libcgpt-cc.a. It doesn't seem to affect the x86 platform. I've started a separate thread with some folks to see if that's the right thing to do.

Patch-set: 6