Add tests/run_preamble_tests.sh and associated data.
This change prepares for modifying VbFirmwarePreambleHeader and
VbKernelPreambleHeader by adding a bunch of current-version data and tests
of that data. Once we change the structs, we'll still need to be sure that
we can still generate, sign, and verify things using the old-style structs
too so that we can release updates to existing devices.

If we changed the structs and then created the test data, we couldn't be
certain that we're still doing it right.

BUG=chromium-os:20124
TEST=manual

make
make runtests

Change-Id: I39310a0d853dbf63a8ca8ff9a0fb4440017c692a