WIP: Switch vboot preamble structs to v3.0
This is work in progress, not to be committed yet. I'm just looking for
confirmation that I haven't screwed anything up in the security realm.

This change does the following:

1. Change VbFirmwarePreambleHeader and VbKernelPreambleHeader to version
   3.0, which uses a hash of the body data for verification instead of
   another signature. Because the preamble header itself is signed,

2. Modify vbutil_firmware and vbutil_kernel to use the new format.

3. Tweak the existing tests as needed to run with the new preamble headers.
   Note that a couple of tests that are specific to version 2.x are
   temporarily skipped.

This change BREAKS vbutil_firmware and vbutil_kernel for pre-3.0 versions.
I'm going to add backwards compatibility to them and re-enable the skipped
version 2.x tests once I'm sure the 3.0 version is implemented correctly.

BUG=chromium-os:20124
TEST=manual

make
make runtests

Change-Id: I1a1c2dab9908ef5b83a07262acf09ce69136fdee