Update patch set 1
Patch Set 1: (1 inline comment)

Do you have a preference for "hash" vs "digest"? Most of the online docs I've read seem to use them interchangeably. Is there a subtle difference that I should be aware of?

Also, would it be less confusing to keep the struct member name "body_signature" instead of changing it to "body_hash", even though it's now containing a digest instead of an encrypted digest?

Patch-set: 1