Update patch set 1
Patch Set 1:

Oh, sorry, i wasn't clear: I agree that there is something wrong with my CL  - it's emitting an error, but I can't reproduce it yet.

Apparently there's something wrong with my workspace too, which tends to happen about once a month or so. 

I'll have to try again in a clean checkout.

Thanks for the help.

Patch-set: 1