INCOMPATIBLE CHANGE to vboot preamble headers
This removes one RSA operation from both firmware and kernel verification,
which makes it significantly faster on some systems. Unfortunately, it means
that systems with RO firmware which expects the old (v2) preamble headers
*WILL NOT ACCEPT* images signed with the new (v3) format.

New systems should use this format and everything should Just Work.

Images for shipping systems must be signed using the old format. Both
vbutil_firmware and vbutil_kernel now include a --format argument to
override the default (new) version although for the kernel, the --repack
option will preserve the format used for the kernel headers. Unfortunately,
there is not a repack function for vbutil_firmware, so re-signing the
firmware shellball will require special care to maintain backwards
compatibility.

For continued support of existing devices, you should use the signing tools
from the release branch, or be SURE to use the --format 2 option. If you
backport this commit to those branches, you may want to modify the default
format value in host_common.h. For top-of-trunk, the use of a per-target
config file may be more convenient. Refer to the comments for the
PreloadOptions() function in host_misc.c, which is used by both signing
utilities.

BUG=chromium-os:20124
TEST=manual

make
make runtests

Change-Id: Idb3841ac1c4a82cf81753041779f02147ede0618