Update patch set 1
Patch Set 1: Do not submit

Hi guys,

As I was told, the kernel body load address check was actually not necessary. And it was a great news to me since in exynos systems, memory address starts from 0x40000000 in U-Boot and CROS_32BIT_ENTRY_ADDR is 0x100000.

However, the interface exposed through vboot_api.h did not seem to provide a way to pass the BOOT_FLAG_SKIP_ADDR_CHECK flag down to LoadKernel(). Did I miss anything in vboot_api.h?

If not, what is the recommended why to suppress the address check? This CL is definitely not the only way!

Patch-set: 1
Label: Code-Review=-2
Label: Commit-Queue=0
Label: Verified=0